refactor: Header 삭제 후 교체
