--- src/components/Common/Header/Header.stories.tsx
@@ -1,13 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import Header from './Header';
-
-const meta: Meta<typeof Header> = {
-  title: 'common/Header',
-  component: Header,
-};
-
-export default meta;
-type Story = StoryObj<typeof Header>;
-
-export const Default: Story = {};
