--- src/components/Common/Header/header.css.ts
@@ -1,23 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const container = style({
-  position: 'fixed',
-  top: 0,
-  left: '50%',
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  width: '100%',
-  height: 50,
-  maxWidth: 440,
-  padding: '0 20px',
-  backgroundColor: '#fff',
-  transform: 'translateX(-50%)',
-  zIndex: 1001,
-});
-
-export const link = style({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
