+++ src/components/Common/TopBar/topBar.css.ts
@@ -11,12 +11,19 @@ export const container = style({
   display: 'flex',
   justifyContent: 'space-between',
   alignItems: 'center',
+  padding: '0 20px',
   backgroundColor: vars.colors.white,
   transform: 'translateX(-50%)',
   zIndex: 1001,
 });
 
-export const LeftNavigationWrapper = style({
+export const link = style({
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+});
+
+export const leftNavigationWrapper = style({
   display: 'flex',
   alignItems: 'center',
 });
